requirements.txt
tensorflow
keras
matplotlib
numpy
pandas
opencv-python
scikit-learn
flask
Pillow
